+++ activemodel/test/cases/validations_test.rb
@@ -237,14 +237,14 @@ class ValidationsTest < ActiveModel::TestCase
     t = Topic.new title: ""
     assert_predicate t, :invalid?
 
-    assert_equal :title, key = assert_deprecated { t.errors.keys[0] }
+    assert_equal :title, key = t.errors.attribute_names[0]
     assert_equal "can't be blank", t.errors[key][0]
     assert_equal "is too short (minimum is 2 characters)", t.errors[key][1]
-    assert_equal :author_name, key = assert_deprecated { t.errors.keys[1] }
+    assert_equal :author_name, key = t.errors.attribute_names[1]
     assert_equal "can't be blank", t.errors[key][0]
-    assert_equal :author_email_address, key = assert_deprecated { t.errors.keys[2] }
+    assert_equal :author_email_address, key = t.errors.attribute_names[2]
     assert_equal "will never be valid", t.errors[key][0]
-    assert_equal :content, key = assert_deprecated { t.errors.keys[3] }
+    assert_equal :content, key = t.errors.attribute_names[3]
     assert_equal "is too short (minimum is 2 characters)", t.errors[key][0]
   end